requirements.txt
langchain
streamlit
python-dotenv
requests
google-genai
google-generativeai